requirements.txt
asgiref==3.8.1
attrs==25.1.0
autopep8==2.3.2
Django==5.1.6
django-environ==0.12.0
django-filter==25.1
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gprof2dot==2024.6.6
inflection==0.5.1
isort==6.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg==3.2.5
psycopg-binary==3.2.5
pycodestyle==2.12.1
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
django-cors-headers==4.7.0
# Python 3.13.0